pyproject.toml
[project]
name = "ai-eng-pi-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langfuse>=4.0.0",
    "python-dotenv>=1.2.2",
]
